+++ src/main/java/org/truffleruby/core/hash/HashNodes.java
@@ -501,7 +501,8 @@ public abstract class HashNodes {
 
         private Object yieldPair(RubyProc block, Object key, Object value) {
             // MRI behavior, see rb_hash_each_pair()
-            if (arityMoreThanOne.profile(block.sharedMethodInfo.getArity().getArityNumber() > 1)) {
+            // We use getMethodArityNumber() here since for non-lambda the semantics are the same for both branches
+            if (arityMoreThanOne.profile(block.sharedMethodInfo.getArity().getMethodArityNumber() > 1)) {
                 return yield(block, key, value);
             } else {
                 return yield(block, createArray(new Object[]{ key, value }));
@@ -722,7 +723,8 @@ public abstract class HashNodes {
 
         private Object yieldPair(RubyProc block, Object key, Object value) {
             // MRI behavior, see rb_hash_each_pair()
-            if (arityMoreThanOne.profile(block.sharedMethodInfo.getArity().getArityNumber() > 1)) {
+            // We use getMethodArityNumber() here since for non-lambda the semantics are the same for both branches
+            if (arityMoreThanOne.profile(block.sharedMethodInfo.getArity().getMethodArityNumber() > 1)) {
                 return yield(block, key, value);
             } else {
                 return yield(block, createArray(new Object[]{ key, value }));